+++ app/spike/internal/cmd/policy/create.go
@@ -13,6 +13,7 @@ import (
 	spike "github.com/spiffe/spike-sdk-go/api"
 	"github.com/spiffe/spike-sdk-go/api/entity/data"
 
+	"github.com/spiffe/spike/app/spike/internal/stdout"
 	"github.com/spiffe/spike/app/spike/internal/trust"
 )
 
@@ -85,11 +86,7 @@ func newPolicyCreateCommand(
 			err := api.CreatePolicy(name, spiffeIddPattern, pathPattern, perms)
 			if err != nil {
 				if err.Error() == "not ready" {
-					fmt.Println("!")
-					fmt.Println("!  SPIKE is not initialized")
-					fmt.Println("!  Wait a few seconds and try again.")
-					fmt.Println("!  If the problem persists, check out SPIKE Nexus logs.")
-					fmt.Println("!")
+					stdout.PrintNotReady()
 					return
 				}