+++ app/spike/internal/cmd/policy/delete.go
@@ -11,6 +11,7 @@ import (
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
 	spike "github.com/spiffe/spike-sdk-go/api"
 
+	"github.com/spiffe/spike/app/spike/internal/stdout"
 	"github.com/spiffe/spike/app/spike/internal/trust"
 )
 
@@ -68,11 +69,7 @@ func newPolicyDeleteCommand(
 			err := api.DeletePolicy(policyID)
 			if err != nil {
 				if err.Error() == "not ready" {
-					fmt.Println("!")
-					fmt.Println("!  SPIKE is not initialized")
-					fmt.Println("!  Wait a few seconds and try again.")
-					fmt.Println("!  If the problem persists, check out SPIKE Nexus logs.")
-					fmt.Println("!")
+					stdout.PrintNotReady()
 					return
 				}