+++ app/spike/internal/cmd/policy/list.go
@@ -10,8 +10,9 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
-
 	spike "github.com/spiffe/spike-sdk-go/api"
+
+	"github.com/spiffe/spike/app/spike/internal/stdout"
 	"github.com/spiffe/spike/app/spike/internal/trust"
 )
 
@@ -87,11 +88,7 @@ func newPolicyListCommand(
 			policies, err := api.ListPolicies()
 			if err != nil {
 				if err.Error() == "not ready" {
-					fmt.Println("!")
-					fmt.Println("!  SPIKE is not initialized")
-					fmt.Println("!  Wait a few seconds and try again.")
-					fmt.Println("!  If the problem persists, check out SPIKE Nexus logs.")
-					fmt.Println("!")
+					stdout.PrintNotReady()
 					return
 				}