+++ app/spike/internal/cmd/secret/metadata_get.go
@@ -9,8 +9,9 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
-
 	spike "github.com/spiffe/spike-sdk-go/api"
+
+	"github.com/spiffe/spike/app/spike/internal/stdout"
 	"github.com/spiffe/spike/app/spike/internal/trust"
 )
 
@@ -63,11 +64,7 @@ func newSecretMetadataGetCommand(
 			secret, err := api.GetSecretMetadata(path, version)
 			if err != nil {
 				if err.Error() == "not ready" {
-					fmt.Println("!")
-					fmt.Println("!  SPIKE is not initialized")
-					fmt.Println("!  Wait a few seconds and try again.")
-					fmt.Println("!  If the problem persists, check out SPIKE Nexus logs.")
-					fmt.Println("!")
+					stdout.PrintNotReady()
 					return
 				}