+++ hack/spire-server-entry-su-register.sh
@@ -31,6 +31,11 @@ if [[ -n $ENTRY_ID ]]; then
       -selector unix:path:"$PILOT_PATH" \
       -selector unix:sha256:"$PILOT_SHA"
 
+  # Wait for the entry to be updated
+  echo "Waiting for entries to be updated..."
+  sleep 5
+  echo "Everything is awesome!"
+
   exit 0
 fi
 
@@ -46,6 +51,11 @@ if [[ -n $ENTRY_ID ]]; then
       -selector unix:uid:"$(id -u)" \
       -selector unix:path:"$PILOT_PATH" \
       -selector unix:sha256:"$PILOT_SHA"
+
+  # Wait for the entry to be updated
+  echo "Waiting for entries to be updated..."
+  sleep 5
+  echo "Everything is awesome!"
 
   exit 0
 fi